server/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.34
botocore==1.35.34
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-formtools==2.5.1
django-oauth-toolkit==2.4.0
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-storages==1.14.4
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.6
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.7
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==7.0.3
minio==7.1.13
